feat: 잘못된 요청이 들어왔을 때 예외 처리 - 여우

+++ src/test/java/racingcar/domain/CarsTest.java
@@ -43,7 +43,7 @@ class CarsTest {
 
         cars.move();
 
-        Long position = car.getPosition();
+        int position = car.getPosition();
         assertThat(position).isEqualTo(1L);
     }
